d2_client/requirements.txt
Django==2.0.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-redis==4.10.0
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
itsdangerous==1.1.0
Markdown==3.1.1
mongoengine==0.18.2
numpy==1.17.4
pandas==0.25.3
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
simplejson==3.16.0
six==1.13.0
